% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.example.r
\docType{data}
\name{data.example}
\alias{data.example}
\title{Example data to illustrate the functions}
\format{
A data frame with 200 observations on the following 5 variables.
  \describe{
    \item{\code{sex}}{\code{"male"} or \code{"female"}}
    \item{\code{age}}{age in days}
    \item{\code{entry_date}}{entry date in the study}
    \item{\code{status}}{status at follow-up time: 0 if alive, 1 if dead}
    \item{\code{futime}}{follow-up time in days}
  }
}
\description{
Example data to illustrate the functions
}
\keyword{data}
