% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survexp.fr-package.r
\docType{package}
\name{survexp.fr-package}
\alias{survexp.fr-package}
\title{Relative survival, AER and SMR based on French death rates}
\description{
Relative survival, AER and SMR based on French death rates
}
\author{
Jean-Philippe Jais and Hugo Varet
}
