\name{survexp.fr}
\alias{data.survexp.fr}
\docType{data}
\title{French data for the expected survival and person years functions}
\description{French data for the expected survival and person years functions}

\usage{data(survexp.fr)}

\details{Death rates are available from 1977 to 2010 for males and females aged from 0 to 99}

\source{\url{http://www.insee.fr/fr/themes/detail.asp?ref_id=ir-sd2010&page=irweb/sd2010/dd/sd2010_mortalite.htm}}

\references{Institut National de la Statistique et des Etudes Economiques}

\examples{
is.ratetable(survexp.fr)
summary(survexp.fr)
}

\keyword{datasets}
