\name{survexp.fr-package}
\alias{survexp.fr-package}
\alias{survexp.fr}
\docType{package}
\title{Relative survival, AER and SMR based on French death rates}
\description{Relative survival, AER and SMR based on French death rates}
\details{
\tabular{ll}{
Package: \tab survexp.fr\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-03-04\cr
License: \tab GPL (>= 2)\cr
}

}
\author{Jean-Philippe Jais and Hugo Varet

Maintainer: Hugo Varet <varethugo@gmail.com>
}


\keyword{ package }

