% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survexp.fr.r
\docType{data}
\name{survexp.fr}
\alias{survexp.fr}
\title{French data for the expected survival and person years functions}
\source{
\url{https://www.insee.fr/fr/statistiques/fichier/5390366/fm_t68.xlsx}
}
\description{
French data for the expected survival and person years functions
}
\details{
Death rates are available from 1977 to 2019 for males and females aged from 0 to 99
}
\references{
Institut National de la Statistique et des Etudes Economiques
}
\keyword{data}
